requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
dm-tree==0.1.5
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.3.0
idna==2.10
importlib-metadata==4.8.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.3
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
Pillow==8.3.2
pip==21.2.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.2
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pyu2f==0.1.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==57.4.0
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-model-optimization==0.5.1.dev0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.35.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
